Move over and slightly refactor integration tests
