+++ athena-mysql/src/test/java/com/amazonaws/athena/connectors/mysql/integ/MySqlIntegTest.java
@@ -1,8 +1,8 @@
 /*-
  * #%L
- * athena-jdbc
+ * athena-mysql
  * %%
- * Copyright (C) 2019 - 2020 Amazon Web Services
+ * Copyright (C) 2019 - 2021 Amazon Web Services
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,19 +17,22 @@
  * limitations under the License.
  * #L%
  */
-package com.amazonaws.athena.connectors.jdbc.integ;
+package com.amazonaws.athena.connectors.mysql.integ;
 
+import com.amazonaws.athena.connector.integ.IntegrationTestBase;
 import com.amazonaws.athena.connector.integ.clients.CloudFormationClient;
 import com.amazonaws.athena.connector.integ.data.ConnectorVpcAttributes;
-import com.amazonaws.athena.connector.integ.IntegrationTestBase;
 import com.amazonaws.athena.connector.integ.data.SecretsManagerCredentials;
 import com.amazonaws.athena.connector.lambda.domain.TableName;
+import com.amazonaws.athena.connectors.jdbc.connection.DatabaseConnectionInfo;
+import com.amazonaws.athena.connectors.jdbc.integ.JdbcTableUtils;
 import com.amazonaws.services.athena.model.Row;
-import com.amazonaws.services.redshift.AmazonRedshift;
-import com.amazonaws.services.redshift.AmazonRedshiftClientBuilder;
-import com.amazonaws.services.redshift.model.DescribeClustersRequest;
-import com.amazonaws.services.redshift.model.DescribeClustersResult;
-import com.amazonaws.services.redshift.model.Endpoint;
+import com.amazonaws.services.rds.AmazonRDS;
+import com.amazonaws.services.rds.AmazonRDSClientBuilder;
+import com.amazonaws.services.rds.model.DescribeDBInstancesRequest;
+import com.amazonaws.services.rds.model.DescribeDBInstancesResult;
+import com.amazonaws.services.rds.model.Endpoint;
+import com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterClass;
@@ -37,16 +40,19 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import software.amazon.awscdk.core.App;
 import software.amazon.awscdk.core.RemovalPolicy;
-import software.amazon.awscdk.core.SecretValue;
 import software.amazon.awscdk.core.Stack;
+import software.amazon.awscdk.services.ec2.InstanceType;
 import software.amazon.awscdk.services.ec2.SecurityGroup;
 import software.amazon.awscdk.services.ec2.Vpc;
 import software.amazon.awscdk.services.ec2.VpcAttributes;
 import software.amazon.awscdk.services.iam.PolicyDocument;
-import software.amazon.awscdk.services.redshift.Cluster;
-import software.amazon.awscdk.services.redshift.ClusterType;
-import software.amazon.awscdk.services.redshift.Login;
-import software.amazon.awscdk.services.redshift.NodeType;
+import software.amazon.awscdk.services.rds.Credentials;
+import software.amazon.awscdk.services.rds.DatabaseInstance;
+import software.amazon.awscdk.services.rds.DatabaseInstanceEngine;
+import software.amazon.awscdk.services.rds.MySqlInstanceEngineProps;
+import software.amazon.awscdk.services.rds.MysqlEngineVersion;
+import software.amazon.awscdk.services.rds.StorageType;
+import software.amazon.awscdk.services.secretsmanager.Secret;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -56,63 +62,69 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
+import static com.amazonaws.athena.connectors.mysql.MySqlConstants.MYSQL_DRIVER_CLASS;
+import static com.amazonaws.athena.connectors.mysql.MySqlConstants.MYSQL_NAME;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 /**
- * Integration-tests for the Redshift (JDBC) connector using the Integration-test module.
+ * Integration-tests for the MySql (JDBC) connector using the Integration-test module.
  */
-public class RedshiftIntegTest extends IntegrationTestBase
+public class MySqlIntegTest extends IntegrationTestBase
 {
-    private static final Logger logger = LoggerFactory.getLogger(RedshiftIntegTest.class);
-
-    private static final long sleepDelayMillis = 60_000L;
+    private static final Logger logger = LoggerFactory.getLogger(MySqlIntegTest.class);
 
     private final App theApp;
+    private final String secretArn;
     private final String username;
     private final String password;
-    private final String redshiftDbName;
-    private final String redshiftDbPort;
-    private final String redshiftTableMovies;
-    private final String redshiftTableBday;
+    private final String mysqlDbName;
+    private final Number mysqlDbPort;
+    private final String mysqlTableMovies;
+    private final String mysqlTableBday;
     private final String lambdaFunctionName;
-    private final String clusterName;
+    private final String dbInstanceName;
     private final Map<String, String> environmentVars;
+    private final Map<String, String> jdbcProperties;
+    private final DatabaseConnectionInfo databaseConnectionInfo;
 
     private CloudFormationClient cloudFormationClient;
 
-    public RedshiftIntegTest()
+    public MySqlIntegTest()
     {
         theApp = new App();
         SecretsManagerCredentials secretsManagerCredentials = getSecretCredentials().orElseThrow(() ->
                 new RuntimeException("secrets_manager_secret must be provided in test-config.json file."));
+        secretArn = secretsManagerCredentials.getArn();
         username = secretsManagerCredentials.getUsername();
         password = secretsManagerCredentials.getPassword();
         Map<String, Object> userSettings = getUserSettings().orElseThrow(() ->
                 new RuntimeException("user_settings attribute must be provided in test-config.json file."));
-        redshiftDbName = (String) userSettings.get("redshift_db_name");
-        redshiftDbPort = (String) userSettings.get("redshift_db_port");
-        redshiftTableMovies = (String) userSettings.get("redshift_table_movies");
-        redshiftTableBday = (String) userSettings.get("redshift_table_bday");
+        mysqlDbName = (String) userSettings.get("mysql_db_name");
+        mysqlDbPort = (Number) userSettings.get("mysql_db_port");
+        mysqlTableMovies = (String) userSettings.get("mysql_table_movies");
+        mysqlTableBday = (String) userSettings.get("mysql_table_bday");
         lambdaFunctionName = getLambdaFunctionName();
-        clusterName = "integ-redshift-cluster-" + UUID.randomUUID();
+        dbInstanceName = "integ-mysql-instance-" + UUID.randomUUID();
         environmentVars = new HashMap<>();
+        jdbcProperties = ImmutableMap.of("databaseTerm", "SCHEMA");
+        databaseConnectionInfo = new DatabaseConnectionInfo(MYSQL_DRIVER_CLASS, (Integer) mysqlDbPort);
     }
 
     /**
-     * Creates a Redshift cluster used for the integration tests.
+     * Creates a MySql RDS Instance used for the integration tests.
      */
     @BeforeClass
     @Override
     protected void setUp()
     {
-        cloudFormationClient = new CloudFormationClient(theApp, getRedshiftStack());
+        cloudFormationClient = new CloudFormationClient(theApp, getMySqlStack());
         try {
-            // Create the CloudFormation stack for the Redshift cluster.
+            // Create the CloudFormation stack for the MySql DB instance.
             cloudFormationClient.createStack();
-            // Get DB cluster's host and port information and set the environment variables needed for the Lambda.
-            setEnvironmentVars(getClusterData());
-            // Create the DB schema in the newly created DB cluster used for the integration tests.
+            // Get DB instance's host and port information and set the environment variables needed for the Lambda.
+            setEnvironmentVars(getInstanceData());
+            // Create the DB schema in the newly created DB instance used for the integration tests.
             createDbSchema();
             // Invoke the framework's setUp().
             super.setUp();
@@ -125,7 +137,7 @@ public class RedshiftIntegTest extends IntegrationTestBase
     }
 
     /**
-     * Deletes a CloudFormation stack for the Redshift cluster.
+     * Deletes a CloudFormation stack for the MySql RDS Instance.
      */
     @AfterClass
     @Override
@@ -133,61 +145,64 @@ public class RedshiftIntegTest extends IntegrationTestBase
     {
         // Invoke the framework's cleanUp().
         super.cleanUp();
-        // Delete the CloudFormation stack for the Redshift cluster.
+        // Delete the CloudFormation stack for the MySql DB instance.
         cloudFormationClient.deleteStack();
     }
 
     /**
-     * Gets the CloudFormation stack for the Redshift cluster.
-     * @return Stack object for the Redshift cluster.
+     * Gets the CloudFormation stack for the MySql RDS Instance.
+     * @return Stack object for the MySql RDS Instance.
      */
-    private Stack getRedshiftStack()
+    private Stack getMySqlStack()
     {
-        Stack stack = Stack.Builder.create(theApp, clusterName).build();
+        Stack stack = Stack.Builder.create(theApp, dbInstanceName).build();
 
         ConnectorVpcAttributes vpcAttributes = getVpcAttributes()
                 .orElseThrow(() -> new RuntimeException("vpc_configuration must be specified in test-config.json"));
 
-        Cluster.Builder.create(stack, "RedshiftCluster")
+        DatabaseInstance.Builder.create(stack, "MySqlInstance")
                 .publiclyAccessible(Boolean.TRUE)
                 .removalPolicy(RemovalPolicy.DESTROY)
-                .encrypted(Boolean.FALSE)
-                .port(Integer.parseInt(redshiftDbPort))
-                .clusterName(clusterName)
-                .clusterType(ClusterType.SINGLE_NODE)
-                .nodeType(NodeType.DC2_LARGE)
-                .numberOfNodes(1)
-                .defaultDatabaseName("public")
-                .masterUser(Login.builder()
-                        .masterUsername(username)
-                        .masterPassword(SecretValue.plainText(password))
-                        .build())
-                .vpc(Vpc.fromVpcAttributes(stack, "RedshiftVpcConfig", VpcAttributes.builder()
+                .deleteAutomatedBackups(Boolean.TRUE)
+                .storageEncrypted(Boolean.FALSE)
+                .port(mysqlDbPort)
+                .instanceIdentifier(dbInstanceName)
+                .engine(DatabaseInstanceEngine.mysql(MySqlInstanceEngineProps.builder()
+                        .version(MysqlEngineVersion.VER_8_0_20)
+                        .build()))
+                .storageType(StorageType.GP2)
+                .allocatedStorage(20)
+                .instanceType(new InstanceType("t2.micro"))
+                .credentials(Credentials.fromSecret(Secret
+                        .fromSecretCompleteArn(stack, "MySqlSecret", secretArn)))
+                .vpc(Vpc.fromVpcAttributes(stack, "MySqlVpcConfig", VpcAttributes.builder()
                         .vpcId(vpcAttributes.getVpcId())
                         .privateSubnetIds(vpcAttributes.getPrivateSubnetIds())
                         .availabilityZones(vpcAttributes.getAvailabilityZones())
                         .build()))
                 .securityGroups(Collections.singletonList(SecurityGroup
-                        .fromSecurityGroupId(stack, "RedshiftVpcSecurityGroup", vpcAttributes.getSecurityGroupId())))
+                        .fromSecurityGroupId(stack, "MySqlVpcSecurityGroup",
+                                vpcAttributes.getSecurityGroupId())))
                 .build();
 
         return stack;
     }
 
     /**
-     * Gets the Redshift cluster endpoint information needed for the Lambda.
-     * All exceptions thrown here will be caught in the calling function.
+     * Gets the MySql RDS Instance endpoint information. All exceptions thrown here will be caught in the
+     * calling function.
+     * @return Endpoint object containing the DB instance's domain and port information.
      */
-    private Endpoint getClusterData()
+    private Endpoint getInstanceData()
     {
-        AmazonRedshift redshiftClient = AmazonRedshiftClientBuilder.defaultClient();
+        AmazonRDS rdsClient = AmazonRDSClientBuilder.defaultClient();
         try {
-            DescribeClustersResult clustersResult = redshiftClient.describeClusters(new DescribeClustersRequest()
-                    .withClusterIdentifier(clusterName));
-            return clustersResult.getClusters().get(0).getEndpoint();
+            DescribeDBInstancesResult instancesResult = rdsClient.describeDBInstances(new DescribeDBInstancesRequest()
+                    .withDBInstanceIdentifier(dbInstanceName));
+            return instancesResult.getDBInstances().get(0).getEndpoint();
         }
         finally {
-            redshiftClient.shutdown();
+            rdsClient.shutdown();
         }
     }
 
@@ -197,7 +212,7 @@ public class RedshiftIntegTest extends IntegrationTestBase
      */
     private void setEnvironmentVars(Endpoint endpoint)
     {
-        String connectionString = String.format("redshift://jdbc:redshift://%s:%s/public?user=%s&password=%s",
+        String connectionString = String.format("mysql://jdbc:mysql://%s:%s/mysql?user=%s&password=%s",
                 endpoint.getAddress(), endpoint.getPort(), username, password);
         String connectionStringTag = lambdaFunctionName + "_connection_string";
         environmentVars.put("default", connectionString);
@@ -210,13 +225,11 @@ public class RedshiftIntegTest extends IntegrationTestBase
     private void createDbSchema()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB Schema: {}", redshiftDbName);
+        logger.info("Setting up DB Schema: {}", mysqlDbName);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils jdbcUtils =
-                new JdbcTableUtils(lambdaFunctionName, new TableName(redshiftDbName, redshiftTableMovies),
-                        environmentVars);
-        jdbcUtils.createDbSchema();
+        JdbcTableUtils jdbcUtils = new JdbcTableUtils(lambdaFunctionName, new TableName(mysqlDbName, mysqlTableMovies), environmentVars, jdbcProperties, MYSQL_NAME);
+        jdbcUtils.createDbSchema(databaseConnectionInfo);
     }
 
     /**
@@ -256,13 +269,6 @@ public class RedshiftIntegTest extends IntegrationTestBase
     @Override
     protected void setUpTableData()
     {
-        try {
-            logger.info("Allowing Redshift cluster to fully warm up - Sleeping for 1 min...");
-            Thread.sleep(sleepDelayMillis);
-        }
-        catch (InterruptedException e) {
-            throw new RuntimeException("Thread.sleep interrupted: " + e.getMessage(), e);
-        }
         setUpMoviesTable();
         setUpBdayTable();
     }
@@ -273,15 +279,13 @@ public class RedshiftIntegTest extends IntegrationTestBase
     private void setUpMoviesTable()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB table: {}", redshiftTableMovies);
+        logger.info("Setting up DB table: {}", mysqlTableMovies);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils moviesTable = new JdbcTableUtils(lambdaFunctionName,
-                new TableName(redshiftDbName, redshiftTableMovies), environmentVars);
-        moviesTable.createTable("year int, title varchar, director varchar, lead varchar");
-        moviesTable.insertRow("2014, 'Interstellar', 'Christopher Nolan', 'Matthew McConaughey'");
-        moviesTable.insertRow("1986, 'Aliens', 'James Cameron', 'Sigourney Weaver'");
-
+        JdbcTableUtils moviesTable = new JdbcTableUtils(lambdaFunctionName, new TableName(mysqlDbName, mysqlTableMovies), environmentVars, jdbcProperties, MYSQL_NAME);
+        moviesTable.createTable("year INTEGER, title VARCHAR(25), director VARCHAR(25), lead_actor VARCHAR(25)", databaseConnectionInfo);
+        moviesTable.insertRow("2014, 'Interstellar', 'Christopher Nolan', 'Matthew McConaughey'", databaseConnectionInfo);
+        moviesTable.insertRow("1986, 'Aliens', 'James Cameron', 'Sigourney Weaver'", databaseConnectionInfo);
     }
 
     /**
@@ -290,15 +294,14 @@ public class RedshiftIntegTest extends IntegrationTestBase
     private void setUpBdayTable()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB table: {}", redshiftTableBday);
+        logger.info("Setting up DB table: {}", mysqlTableBday);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils bdayTable = new JdbcTableUtils(lambdaFunctionName,
-                new TableName(redshiftDbName, redshiftTableBday), environmentVars);
-        bdayTable.createTable("first_name varchar, last_name varchar, birthday date");
-        bdayTable.insertRow("'Joe', 'Schmoe', date('2002-05-05')");
-        bdayTable.insertRow("'Jane', 'Doe', date('2005-10-12')");
-        bdayTable.insertRow("'John', 'Smith', date('2006-02-10')");
+        JdbcTableUtils bdayTable = new JdbcTableUtils(lambdaFunctionName, new TableName(mysqlDbName, mysqlTableBday), environmentVars, jdbcProperties, MYSQL_NAME);
+        bdayTable.createTable("first_name VARCHAR(10), last_name VARCHAR(10), birthday DATE", databaseConnectionInfo);
+        bdayTable.insertRow("'Joe', 'Schmoe', '2002-05-05'", databaseConnectionInfo);
+        bdayTable.insertRow("'Jane', 'Doe', '2005-10-12'", databaseConnectionInfo);
+        bdayTable.insertRow("'John', 'Smith', '2006-02-10'", databaseConnectionInfo);
     }
 
     @Test
@@ -310,7 +313,7 @@ public class RedshiftIntegTest extends IntegrationTestBase
 
         List dbNames = listDatabases();
         logger.info("Databases: {}", dbNames);
-        assertTrue("DB not found.", dbNames.contains(redshiftDbName));
+        assertTrue("DB not found.", dbNames.contains(mysqlDbName));
     }
 
     @Test
@@ -320,13 +323,13 @@ public class RedshiftIntegTest extends IntegrationTestBase
         logger.info("Executing listTablesIntegTest");
         logger.info("-----------------------------------");
 
-        List tableNames = listTables(redshiftDbName);
+        List tableNames = listTables(mysqlDbName);
         logger.info("Tables: {}", tableNames);
         assertEquals("Incorrect number of tables found.", 2, tableNames.size());
-        assertTrue(String.format("Table not found: %s.", redshiftTableMovies),
-                tableNames.contains(redshiftTableMovies));
-        assertTrue(String.format("Table not found: %s.", redshiftTableBday),
-                tableNames.contains(redshiftTableBday));
+        assertTrue(String.format("Table not found: %s.", mysqlTableMovies),
+                tableNames.contains(mysqlTableMovies));
+        assertTrue(String.format("Table not found: %s.", mysqlTableBday),
+                tableNames.contains(mysqlTableBday));
     }
 
     @Test
@@ -336,7 +339,7 @@ public class RedshiftIntegTest extends IntegrationTestBase
         logger.info("Executing listTableSchemaIntegTest");
         logger.info("--------------------------------------");
 
-        Map schema = describeTable(redshiftDbName, redshiftTableMovies);
+        Map schema = describeTable(mysqlDbName, mysqlTableMovies);
         schema.remove("partition_name");
         schema.remove("partition_schema_name");
         logger.info("Schema: {}", schema);
@@ -347,8 +350,8 @@ public class RedshiftIntegTest extends IntegrationTestBase
         assertEquals("Wrong column type for title.", "varchar", schema.get("title"));
         assertTrue("Column not found: director", schema.containsKey("director"));
         assertEquals("Wrong column type for director.", "varchar", schema.get("director"));
-        assertTrue("Column not found: lead", schema.containsKey("lead"));
-        assertEquals("Wrong column type for lead.", "varchar", schema.get("lead"));
+        assertTrue("Column not found: lead_actor", schema.containsKey("lead_actor"));
+        assertEquals("Wrong column type for lead_actor.", "varchar", schema.get("lead_actor"));
     }
 
     @Test
@@ -358,8 +361,8 @@ public class RedshiftIntegTest extends IntegrationTestBase
         logger.info("Executing selectColumnWithPredicateIntegTest");
         logger.info("--------------------------------------------------");
 
-        String query = String.format("select title from %s.%s.%s where year > 2000;",
-                lambdaFunctionName, redshiftDbName, redshiftTableMovies);
+        String query = String.format("select title from %s.%s.%s where year > 2010;",
+                lambdaFunctionName, mysqlDbName, mysqlTableMovies);
         List<Row> rows = startQueryExecution(query).getResultSet().getRows();
         if (!rows.isEmpty()) {
             // Remove the column-header row
@@ -373,15 +376,14 @@ public class RedshiftIntegTest extends IntegrationTestBase
     }
 
     @Test
-    public void selectColumnBetweenDatesIntegTest()
-    {
+    public void selectColumnBetweenDatesIntegTest() {
         logger.info("--------------------------------------------------");
         logger.info("Executing selectColumnBetweenDatesIntegTest");
         logger.info("--------------------------------------------------");
 
         String query = String.format(
-                "select first_name from %s.%s.%s where birthday between date('2003-1-1') and date('2005-12-31');",
-                lambdaFunctionName, redshiftDbName, redshiftTableBday);
+                "select first_name from %s.%s.%s where birthday between date('2005-10-01') and date('2005-10-31');",
+                lambdaFunctionName, mysqlDbName, mysqlTableBday);
         List<Row> rows = startQueryExecution(query).getResultSet().getRows();
         if (!rows.isEmpty()) {
             // Remove the column-header row
