+++ athena-redshift/src/test/java/com/amazonaws/athena/connectors/redshift/integ/RedshiftIntegTest.java
@@ -1,15 +1,15 @@
 /*-
  * #%L
- * athena-jdbc
+ * athena-redshift
  * %%
- * Copyright (C) 2019 - 2021 Amazon Web Services
+ * Copyright (C) 2019 - 2020 Amazon Web Services
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,20 +17,21 @@
  * limitations under the License.
  * #L%
  */
-package com.amazonaws.athena.connectors.jdbc.integ;
+package com.amazonaws.athena.connectors.redshift.integ;
 
 import com.amazonaws.athena.connector.integ.IntegrationTestBase;
 import com.amazonaws.athena.connector.integ.clients.CloudFormationClient;
 import com.amazonaws.athena.connector.integ.data.ConnectorVpcAttributes;
 import com.amazonaws.athena.connector.integ.data.SecretsManagerCredentials;
 import com.amazonaws.athena.connector.lambda.domain.TableName;
+import com.amazonaws.athena.connectors.jdbc.connection.DatabaseConnectionInfo;
+import com.amazonaws.athena.connectors.jdbc.integ.JdbcTableUtils;
 import com.amazonaws.services.athena.model.Row;
-import com.amazonaws.services.rds.AmazonRDS;
-import com.amazonaws.services.rds.AmazonRDSClientBuilder;
-import com.amazonaws.services.rds.model.DescribeDBInstancesRequest;
-import com.amazonaws.services.rds.model.DescribeDBInstancesResult;
-import com.amazonaws.services.rds.model.Endpoint;
-import com.google.common.collect.ImmutableMap;
+import com.amazonaws.services.redshift.AmazonRedshift;
+import com.amazonaws.services.redshift.AmazonRedshiftClientBuilder;
+import com.amazonaws.services.redshift.model.DescribeClustersRequest;
+import com.amazonaws.services.redshift.model.DescribeClustersResult;
+import com.amazonaws.services.redshift.model.Endpoint;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterClass;
@@ -38,19 +39,16 @@ import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import software.amazon.awscdk.core.App;
 import software.amazon.awscdk.core.RemovalPolicy;
+import software.amazon.awscdk.core.SecretValue;
 import software.amazon.awscdk.core.Stack;
-import software.amazon.awscdk.services.ec2.InstanceType;
 import software.amazon.awscdk.services.ec2.SecurityGroup;
 import software.amazon.awscdk.services.ec2.Vpc;
 import software.amazon.awscdk.services.ec2.VpcAttributes;
 import software.amazon.awscdk.services.iam.PolicyDocument;
-import software.amazon.awscdk.services.rds.Credentials;
-import software.amazon.awscdk.services.rds.DatabaseInstance;
-import software.amazon.awscdk.services.rds.DatabaseInstanceEngine;
-import software.amazon.awscdk.services.rds.PostgresInstanceEngineProps;
-import software.amazon.awscdk.services.rds.PostgresEngineVersion;
-import software.amazon.awscdk.services.rds.StorageType;
-import software.amazon.awscdk.services.secretsmanager.Secret;
+import software.amazon.awscdk.services.redshift.Cluster;
+import software.amazon.awscdk.services.redshift.ClusterType;
+import software.amazon.awscdk.services.redshift.Login;
+import software.amazon.awscdk.services.redshift.NodeType;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -60,65 +58,68 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
+import static com.amazonaws.athena.connectors.redshift.RedshiftConstants.REDSHIFT_DEFAULT_PORT;
+import static com.amazonaws.athena.connectors.redshift.RedshiftConstants.REDSHIFT_DRIVER_CLASS;
+import static com.amazonaws.athena.connectors.redshift.RedshiftConstants.REDSHIFT_NAME;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 /**
- * Integration-tests for the PostGreSql (JDBC) connector using the Integration-test module.
+ * Integration-tests for the Redshift (JDBC) connector using the Integration-test module.
  */
-public class PostGreSqlIntegTest extends IntegrationTestBase
+public class RedshiftIntegTest extends IntegrationTestBase
 {
-    private static final Logger logger = LoggerFactory.getLogger(PostGreSqlIntegTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(RedshiftIntegTest.class);
+
+    private static final long sleepDelayMillis = 60_000L;
 
     private final App theApp;
-    private final String secretArn;
     private final String username;
     private final String password;
-    private final String postgresDbName;
-    private final Number postgresDbPort;
-    private final String postgresTableMovies;
-    private final String postgresTableBday;
+    private final String redshiftDbName;
+    private final String redshiftDbPort;
+    private final String redshiftTableMovies;
+    private final String redshiftTableBday;
     private final String lambdaFunctionName;
-    private final String dbInstanceName;
+    private final String clusterName;
     private final Map<String, String> environmentVars;
-    private final Map<String, String> jdbcProperties;
+    private final DatabaseConnectionInfo databaseConnectionInfo;
 
     private CloudFormationClient cloudFormationClient;
 
-    public PostGreSqlIntegTest()
+    public RedshiftIntegTest()
     {
         theApp = new App();
         SecretsManagerCredentials secretsManagerCredentials = getSecretCredentials().orElseThrow(() ->
                 new RuntimeException("secrets_manager_secret must be provided in test-config.json file."));
-        secretArn = secretsManagerCredentials.getArn();
         username = secretsManagerCredentials.getUsername();
         password = secretsManagerCredentials.getPassword();
         Map<String, Object> userSettings = getUserSettings().orElseThrow(() ->
                 new RuntimeException("user_settings attribute must be provided in test-config.json file."));
-        postgresDbName = (String) userSettings.get("postgres_db_name");
-        postgresDbPort = (Number) userSettings.get("postgres_db_port");
-        postgresTableMovies = (String) userSettings.get("postgres_table_movies");
-        postgresTableBday = (String) userSettings.get("postgres_table_bday");
+        redshiftDbName = (String) userSettings.get("redshift_db_name");
+        redshiftDbPort = (String) userSettings.get("redshift_db_port");
+        redshiftTableMovies = (String) userSettings.get("redshift_table_movies");
+        redshiftTableBday = (String) userSettings.get("redshift_table_bday");
         lambdaFunctionName = getLambdaFunctionName();
-        dbInstanceName = "integ-postgres-instance-" + UUID.randomUUID();
+        clusterName = "integ-redshift-cluster-" + UUID.randomUUID();
         environmentVars = new HashMap<>();
-        jdbcProperties = ImmutableMap.of("databaseTerm", "SCHEMA");
+        databaseConnectionInfo = new DatabaseConnectionInfo(REDSHIFT_DRIVER_CLASS, REDSHIFT_DEFAULT_PORT);
     }
 
     /**
-     * Creates a PostGreSql RDS Instance used for the integration tests.
+     * Creates a Redshift cluster used for the integration tests.
      */
     @BeforeClass
     @Override
     protected void setUp()
     {
-        cloudFormationClient = new CloudFormationClient(theApp, getPostGreSqlStack());
+        cloudFormationClient = new CloudFormationClient(theApp, getRedshiftStack());
         try {
-            // Create the CloudFormation stack for the PostGreSql DB instance.
+            // Create the CloudFormation stack for the Redshift cluster.
             cloudFormationClient.createStack();
-            // Get DB instance's host and port information and set the environment variables needed for the Lambda.
-            setEnvironmentVars(getInstanceData());
-            // Create the DB schema in the newly created DB instance used for the integration tests.
+            // Get DB cluster's host and port information and set the environment variables needed for the Lambda.
+            setEnvironmentVars(getClusterData());
+            // Create the DB schema in the newly created DB cluster used for the integration tests.
             createDbSchema();
             // Invoke the framework's setUp().
             super.setUp();
@@ -131,7 +132,7 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     }
 
     /**
-     * Deletes a CloudFormation stack for the PostGreSql RDS Instance.
+     * Deletes a CloudFormation stack for the Redshift cluster.
      */
     @AfterClass
     @Override
@@ -139,63 +140,61 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     {
         // Invoke the framework's cleanUp().
         super.cleanUp();
-        // Delete the CloudFormation stack for the PostGreSql DB instance.
+        // Delete the CloudFormation stack for the Redshift cluster.
         cloudFormationClient.deleteStack();
     }
 
     /**
-     * Gets the CloudFormation stack for the PostGreSql RDS Instance.
-     * @return Stack object for the PostGreSql RDS Instance.
+     * Gets the CloudFormation stack for the Redshift cluster.
+     * @return Stack object for the Redshift cluster.
      */
-    private Stack getPostGreSqlStack()
+    private Stack getRedshiftStack()
     {
-        Stack stack = Stack.Builder.create(theApp, dbInstanceName).build();
+        Stack stack = Stack.Builder.create(theApp, clusterName).build();
 
         ConnectorVpcAttributes vpcAttributes = getVpcAttributes()
                 .orElseThrow(() -> new RuntimeException("vpc_configuration must be specified in test-config.json"));
 
-        DatabaseInstance.Builder.create(stack, "PostGreSqlInstance")
+        Cluster.Builder.create(stack, "RedshiftCluster")
                 .publiclyAccessible(Boolean.TRUE)
                 .removalPolicy(RemovalPolicy.DESTROY)
-                .deleteAutomatedBackups(Boolean.TRUE)
-                .storageEncrypted(Boolean.FALSE)
-                .port(postgresDbPort)
-                .instanceIdentifier(dbInstanceName)
-                .engine(DatabaseInstanceEngine.postgres(PostgresInstanceEngineProps.builder()
-                        .version(PostgresEngineVersion.VER_12_4)
-                        .build()))
-                .storageType(StorageType.GP2)
-                .allocatedStorage(20)
-                .instanceType(new InstanceType("t2.micro"))
-                .credentials(Credentials.fromSecret(Secret
-                        .fromSecretCompleteArn(stack, "PostGreSqlSecret", secretArn)))
-                .vpc(Vpc.fromVpcAttributes(stack, "PostGreSqlVpcConfig", VpcAttributes.builder()
+                .encrypted(Boolean.FALSE)
+                .port(Integer.parseInt(redshiftDbPort))
+                .clusterName(clusterName)
+                .clusterType(ClusterType.SINGLE_NODE)
+                .nodeType(NodeType.DC2_LARGE)
+                .numberOfNodes(1)
+                .defaultDatabaseName("public")
+                .masterUser(Login.builder()
+                        .masterUsername(username)
+                        .masterPassword(SecretValue.plainText(password))
+                        .build())
+                .vpc(Vpc.fromVpcAttributes(stack, "RedshiftVpcConfig", VpcAttributes.builder()
                         .vpcId(vpcAttributes.getVpcId())
                         .privateSubnetIds(vpcAttributes.getPrivateSubnetIds())
                         .availabilityZones(vpcAttributes.getAvailabilityZones())
                         .build()))
                 .securityGroups(Collections.singletonList(SecurityGroup
-                        .fromSecurityGroupId(stack, "PostGreSqlVpcSecurityGroup",
-                                vpcAttributes.getSecurityGroupId())))
+                        .fromSecurityGroupId(stack, "RedshiftVpcSecurityGroup", vpcAttributes.getSecurityGroupId())))
                 .build();
 
         return stack;
     }
 
     /**
-     * Gets the PostGreSql RDS Instance endpoint information and generates the environment variables needed for the
-     * Lambda. All exceptions thrown here will be caught in the calling function.
+     * Gets the Redshift cluster endpoint information needed for the Lambda.
+     * All exceptions thrown here will be caught in the calling function.
      */
-    private Endpoint getInstanceData()
+    private Endpoint getClusterData()
     {
-        AmazonRDS rdsClient = AmazonRDSClientBuilder.defaultClient();
+        AmazonRedshift redshiftClient = AmazonRedshiftClientBuilder.defaultClient();
         try {
-            DescribeDBInstancesResult instancesResult = rdsClient.describeDBInstances(new DescribeDBInstancesRequest()
-                    .withDBInstanceIdentifier(dbInstanceName));
-            return instancesResult.getDBInstances().get(0).getEndpoint();
+            DescribeClustersResult clustersResult = redshiftClient.describeClusters(new DescribeClustersRequest()
+                    .withClusterIdentifier(clusterName));
+            return clustersResult.getClusters().get(0).getEndpoint();
         }
         finally {
-            rdsClient.shutdown();
+            redshiftClient.shutdown();
         }
     }
 
@@ -205,7 +204,7 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
      */
     private void setEnvironmentVars(Endpoint endpoint)
     {
-        String connectionString = String.format("postgres://jdbc:postgresql://%s:%s/postgres?user=%s&password=%s",
+        String connectionString = String.format("redshift://jdbc:redshift://%s:%s/public?user=%s&password=%s",
                 endpoint.getAddress(), endpoint.getPort(), username, password);
         String connectionStringTag = lambdaFunctionName + "_connection_string";
         environmentVars.put("default", connectionString);
@@ -218,13 +217,11 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     private void createDbSchema()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB Schema: {}", postgresDbName);
+        logger.info("Setting up DB Schema: {}", redshiftDbName);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils jdbcUtils =
-                new JdbcTableUtils(lambdaFunctionName, new TableName(postgresDbName, postgresTableMovies),
-                        environmentVars, jdbcProperties);
-        jdbcUtils.createDbSchema();
+        JdbcTableUtils jdbcUtils = new JdbcTableUtils(lambdaFunctionName, new TableName(redshiftDbName, redshiftTableMovies), environmentVars, null, REDSHIFT_NAME);
+        jdbcUtils.createDbSchema(databaseConnectionInfo);
     }
 
     /**
@@ -264,6 +261,13 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     @Override
     protected void setUpTableData()
     {
+        try {
+            logger.info("Allowing Redshift cluster to fully warm up - Sleeping for 1 min...");
+            Thread.sleep(sleepDelayMillis);
+        }
+        catch (InterruptedException e) {
+            throw new RuntimeException("Thread.sleep interrupted: " + e.getMessage(), e);
+        }
         setUpMoviesTable();
         setUpBdayTable();
     }
@@ -274,18 +278,13 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     private void setUpMoviesTable()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB table: {}", postgresTableMovies);
+        logger.info("Setting up DB table: {}", redshiftTableMovies);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils moviesTable =
-                new JdbcTableUtils(lambdaFunctionName, new TableName(postgresDbName, postgresTableMovies),
-                        environmentVars, jdbcProperties);
-        moviesTable.createTable("year int, title varchar, director varchar, actors varchar[]");
-        moviesTable.insertRow("2014, 'Interstellar', 'Christopher Nolan', " +
-                "'{Matthew McConaughey, John Lithgow, Ann Hathaway, David Gyasi, Michael Caine, " +
-                "Jessica Chastain, Matt Damon, Casey Affleck}'");
-        moviesTable.insertRow("1986, 'Aliens', 'James Cameron', " +
-                "'{Sigourney Weaver, Paul Reiser, Lance Henriksen, Bill Paxton}'");
+        JdbcTableUtils moviesTable = new JdbcTableUtils(lambdaFunctionName, new TableName(redshiftDbName, redshiftTableMovies), environmentVars, null, REDSHIFT_NAME);
+        moviesTable.createTable("year int, title varchar, director varchar, lead varchar", databaseConnectionInfo);
+        moviesTable.insertRow("2014, 'Interstellar', 'Christopher Nolan', 'Matthew McConaughey'", databaseConnectionInfo);
+        moviesTable.insertRow("1986, 'Aliens', 'James Cameron', 'Sigourney Weaver'", databaseConnectionInfo);
 
     }
 
@@ -295,16 +294,15 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
     private void setUpBdayTable()
     {
         logger.info("----------------------------------------------------");
-        logger.info("Setting up DB table: {}", postgresTableBday);
+        logger.info("Setting up DB table: {}", redshiftTableBday);
         logger.info("----------------------------------------------------");
 
-        JdbcTableUtils bdayTable =
-                new JdbcTableUtils(lambdaFunctionName, new TableName(postgresDbName, postgresTableBday),
-                        environmentVars, jdbcProperties);
-        bdayTable.createTable("first_name varchar, last_name varchar, birthday date");
-        bdayTable.insertRow("'Joe', 'Schmoe', date('2002-05-05')");
-        bdayTable.insertRow("'Jane', 'Doe', date('2005-10-12')");
-        bdayTable.insertRow("'John', 'Smith', date('2006-02-10')");
+        JdbcTableUtils bdayTable = new JdbcTableUtils(lambdaFunctionName,
+                new TableName(redshiftDbName, redshiftTableBday), environmentVars, null, REDSHIFT_NAME);
+        bdayTable.createTable("first_name varchar, last_name varchar, birthday date", databaseConnectionInfo);
+        bdayTable.insertRow("'Joe', 'Schmoe', date('2002-05-05')", databaseConnectionInfo);
+        bdayTable.insertRow("'Jane', 'Doe', date('2005-10-12')", databaseConnectionInfo);
+        bdayTable.insertRow("'John', 'Smith', date('2006-02-10')", databaseConnectionInfo);
     }
 
     @Test
@@ -316,7 +314,7 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
 
         List dbNames = listDatabases();
         logger.info("Databases: {}", dbNames);
-        assertTrue("DB not found.", dbNames.contains(postgresDbName));
+        assertTrue("DB not found.", dbNames.contains(redshiftDbName));
     }
 
     @Test
@@ -326,13 +324,13 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
         logger.info("Executing listTablesIntegTest");
         logger.info("-----------------------------------");
 
-        List tableNames = listTables(postgresDbName);
+        List tableNames = listTables(redshiftDbName);
         logger.info("Tables: {}", tableNames);
         assertEquals("Incorrect number of tables found.", 2, tableNames.size());
-        assertTrue(String.format("Table not found: %s.", postgresTableMovies),
-                tableNames.contains(postgresTableMovies));
-        assertTrue(String.format("Table not found: %s.", postgresTableBday),
-                tableNames.contains(postgresTableBday));
+        assertTrue(String.format("Table not found: %s.", redshiftTableMovies),
+                tableNames.contains(redshiftTableMovies));
+        assertTrue(String.format("Table not found: %s.", redshiftTableBday),
+                tableNames.contains(redshiftTableBday));
     }
 
     @Test
@@ -342,7 +340,7 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
         logger.info("Executing listTableSchemaIntegTest");
         logger.info("--------------------------------------");
 
-        Map schema = describeTable(postgresDbName, postgresTableMovies);
+        Map schema = describeTable(redshiftDbName, redshiftTableMovies);
         schema.remove("partition_name");
         schema.remove("partition_schema_name");
         logger.info("Schema: {}", schema);
@@ -353,8 +351,8 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
         assertEquals("Wrong column type for title.", "varchar", schema.get("title"));
         assertTrue("Column not found: director", schema.containsKey("director"));
         assertEquals("Wrong column type for director.", "varchar", schema.get("director"));
-        assertTrue("Column not found: actors", schema.containsKey("actors"));
-        assertEquals("Wrong column type for actors.", "array<varchar>", schema.get("actors"));
+        assertTrue("Column not found: lead", schema.containsKey("lead"));
+        assertEquals("Wrong column type for lead.", "varchar", schema.get("lead"));
     }
 
     @Test
@@ -364,8 +362,8 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
         logger.info("Executing selectColumnWithPredicateIntegTest");
         logger.info("--------------------------------------------------");
 
-        String query = String.format("select title from %s.%s.%s where year > 2010;",
-                lambdaFunctionName, postgresDbName, postgresTableMovies);
+        String query = String.format("select title from %s.%s.%s where year > 2000;",
+                lambdaFunctionName, redshiftDbName, redshiftTableMovies);
         List<Row> rows = startQueryExecution(query).getResultSet().getRows();
         if (!rows.isEmpty()) {
             // Remove the column-header row
@@ -386,8 +384,8 @@ public class PostGreSqlIntegTest extends IntegrationTestBase
         logger.info("--------------------------------------------------");
 
         String query = String.format(
-                "select first_name from %s.%s.%s where birthday between date('2005-10-01') and date('2005-10-31');",
-                lambdaFunctionName, postgresDbName, postgresTableBday);
+                "select first_name from %s.%s.%s where birthday between date('2003-1-1') and date('2005-12-31');",
+                lambdaFunctionName, redshiftDbName, redshiftTableBday);
         List<Row> rows = startQueryExecution(query).getResultSet().getRows();
         if (!rows.isEmpty()) {
             // Remove the column-header row
